impliment algo and fix bughs

+++ rules_part_one.c
@@ -6,13 +6,13 @@
 /*   By: del-yaag <del-yaag@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/12 15:24:00 by del-yaag          #+#    #+#             */
-/*   Updated: 2022/12/18 15:39:53 by del-yaag         ###   ########.fr       */
+/*   Updated: 2022/12/21 17:23:29 by del-yaag         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-size_t	ft_lstsize(t_stack *lst)
+int	ft_lstsize(t_stack *lst)
 {
 	size_t		count;
 	t_stack	*ptr;
